requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.27
pymysql==1.1.0
python-dotenv==1.1.1
pydantic==2.11.7
python-multipart==0.0.9
cryptography==42.0.8
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0 